+++ CHANGES.rst
@@ -6,6 +6,7 @@ Changes to be released in next version
 
 🙌 Improvements
  * i18n: Add Estonian (et).
+ * MXSession: Make vc_canSetupSecureBackup reusable.
 
 🐛 Bugfix
  * Settings: New phone number is invisible in dark theme (#3218).